requirements.txt
torch
torchvision
numpy
scipy
librosa
matplotlib
tqdm
Pillow
annoy
